Transform code to be more object oriented
This was kinda necessary to have a clean solution for the loading
problem and the recursive state dependency. This now allows to create
instances and make the window modules completely independent. Though
this could have been still implemented with pure functions and the
objects as tables only. But I wanted to try so object oriented patterns
in Lua the first time.