Add new warning message window
It was is a bit annoying that if no code actions are available, the
warning message only appears at the command line at the very bottom of
the editor. As the user is rather expecting  window to pop-up close to
the cursor, warning messages now open in a window at the exact same
location as the menu would.