Introduce action classes
The term `action` has been introduced to solve issue that the server can
return a list of `Command` or `CodeAction` objects. Therefore many parts
in the plugin have been renamed from `code action` to `action` to
represent that this could be either a `Command` or an actual
`CodeAction` instance.

The former functional modules to execute actions and format them has
been integrated into the action derivation classes.