Fix window positioning for real this time
I finally found the issue. The menu window its location got determined
by the LSP utility function of the NeoVim shared runtime module. When
the cursor is above the half of the editor height, the returned window
position as used for the details window are the actual lines above plus
one. As soon as the cursor moves below the half, it is the exact number
of lines. This changing offset by one cases the issue that the details
window is either positioned correctly or not.
As from the details windows view it is not possible to determine this
and correct it as there is no knowledge about the origin window, the
approach is to do an absolute and manual position of the menu window.
This leads to correct return values of the menu window its position and
thereby the details window works fine.