Close possibly open window instances before open
This is for example the case if the user tries to get code actions,
there are no (warning window pops up) and then without moving the cursor
tries again to get the menu. Might happen when the user expects the
language server to still being in the start-up phase or similar.
In general it is a good idea to make a clean table first.