Export action diff into new window
This window isn't yet fully working as it can't dock correctly at the
bottom of the screen. It will draw behind the menu in this case. This
requires some more complicated improvements.
The future goal is that the diff window will also include fold with the
actual diffs to the files. Not only the summaries.