Convert docking windows into stacking ones
The approach to dock a window was developed when there were only two
windows. It does not work properly for more windows. In fact it lacks
the knowledge in which continuous direction the windows should stack
without hiding each other.
The stack concept directly tackles this issue. Each window must only
know about its predecessor windows in the stack.