Add the concept of an anchor window for stacks
There was the issue of getting the right window stack direction as it
only starts after the first window which is the menu window. It still
got placed manually based on the remaining space according to the cursor
position. So basically the direction gets already decided here. Then the
first window that gets placed in the stack is the details window that
takes only the menu window as reference in the stack to decide on
a direction now. By math, this must be always the revers direction of
the previous one. Thereby it will always hide the current line the code
action comes from.
To solve this issue, also the menu window gets placed by the stack. But
therefore it needs another window in the stack as predecessor, the
anchor. The anchor is in fact as it was before the cursor position.
Therefore the anchor window implements the regular window class, but
is kinda fake as it provides window options according to the cursor.