Add colored squares for diff window
This is a feature the user might be used to from GitHub where the diff
of a PR or file is not only shown by the absolute numbers, but also
5 colored squares which show the proportions of the changes.
The implementation got inspired by the one of Octo.nvim, but got adopted
and tweaked what I works looks better. Octo.nvim is definitely not 100%
as GitHub to my observation. I'm not sure if mine is.