Make buffers being re-used and update the content
Instead of creating a new replacing buffer for an update of the content
in any of the windows, it now re-uses an already existing one.
Therefore the window instance also holds the information about the open
buffer. If none has been created yet, it will open one. Else it just
replaces the content of the buffer. This works basically exactly as the
handling of the windows themselves.
Furthermore this allows for some simplification and removal of
duplicated code among the deriving window classes.

This feature requires the filetype plugins to clean up their local
settings on filetype changes. This is necessary to allow a temporally
removal of any options which would prevent the update of a buffers
content. To keep the window class logic independent of which options
would prevent it, unsetting and re-setting the filetype option again
acts as generic solution here.

Unfortunately this temporally disables the feature of the virtual text for the
diff summary. But this feature got broken anyway on the last commit as it does
not handle the preview of the changing lines in between the summaries. This just
don't happen often, as many code actions include a single text edit only. It is
target of the next commit to fix this bug and bring the feature back.