fix empty diff window
The diff window should be closed (or not even opened) if the currently
selected action has no diff. This was incorrectly implemented, because
the window got closed, but was then always re-opened again.
In case the first action of a new menu has no diff window and the user never
selects an action with a diff, this will also lead to the bug that the next time
the menu opens, the old (empty) diff buffer still exists because it got never
displayed and thereby the wipe option doesn't take action. As resulting issue
the next menu will fail as it tries to create a buffer with an already in use
buffer name.