improve the window width of the stack
So far each window in the stack was using the same width as the first
one (except for the anchors child). This comes with the big downside
that if the first window (i.e. the menu) is very short all following
windows might not fit their content and have ugly line wraps.
To fix this, the whole window stack always has the width of the widest
buffer/window.