enforce buffer deletion when window closes
Unfortunately it seems to happen to some users that the menu closes, but
the buffers are not wiped correctly. As result there will be an error
because a buffer with the same name does already exist. This tries to
improve it by manually deleting the buffer when the window closes.
Anyway I'll keep the wipe option for the moment, even it is kinda
duplicated. I hope both together should make this safe.

I might need to remove the buffer naming if I can't get this under
control. The naming was (if I remember correctly) more a nice to have
thing to have no unlisted buffers the user doesn't know what it is. So
it is not necessary for functionality.