add check that home directory path is defined
Apparently under Windows, the HOME environment variable could not be
detected correctly (in some cases?). In result the variable is `nil` and
leads to an error when calling the `find` function with it. To make the
plugin more stable, I added an extra check for `nil` first. In this case
it will fall back to show absolute paths as always.