remove buffer name to fix duplicated buffer issues
Apparently it is not possible to get 100% rid of the duplicated buffer
issue. Despite all buffers by this plugin should be wiped automatically
and also deleted manually, users still have issues. There I'll now
remove the buffer naming to prevent this issue. The former mechanisms to
clear the buffers remain in place at the buffers should not accumulate
over time. But if it happens that a buffer remains existing, it will not
interfere anymore with the new one. Due to the over buffer options this
should cause no issues to the user despite an unlisted buffer.