add ability to resolve code actions
On the latest nightly of NeoVim it got added the LSP client ability to
resolve code actions. As result some servers now send code actions that
need to be resolved. As result these actions get interpreted wrongly and
the final execution fails.
To solve this, it will be checked first if this actions needs to be
resolved and do so if that is the case. Only afterwards the data gets
interpreted.

Furthermore the interpretation of the data the server sends has become
more robust. All cases are now strict and there is no fallback case
anymore as before. If it fails to interpret the data, it will warn the
user with a notification but continue with the rest of the actions it
can understand.