improve parsing of server responses with errors
So far I was kinda stupid as I simply forgot to handle the case that the
server might respond with an error. It has been discovered while
working with resolvable code actions.
This has been solved now. While doing so, it was required to refactor
the functions for the action request process a little and handle
language clients/servers more individually. This decision was made to
have unified handling for the response of a single server. Before it was
a mix of either the response of a single server (code action resolve
request) or multiple servers (get code actions). Unfortunately this
requires some additional work, but the code structure should be better.