remove special highlighting for warning messages
The warning message filetype plugin sets a special highlighting its floating
window border. Therefore it uses the `winhighlight` option to make this change
locally. This feature of NeoVim does not work very well and already caused some
issues in the past. Now it also breaks the floating window border highlight of
users after the warning message window closes.
As this is by all means not a necessary feature, I just remove it for the
moment. I'll have to research how other plugins like `vim-notify` do this
properly and then re-add this feature. Hopefully in a more advanced manner with
more options for the user for all windows of this plugin.