add configuration to disable parts of the UI
Some users do not like to see parts of the UI. While the diff window will not be
displayed if there is nothing to preview, the details window always pops up. To
allow for better customization, global variables have been added to control the
visibility of each window, except the main menu.