Fix getting actions from all clients
The table of clients may be sparse and have empty slots. This can happen
when LSP clients are restarted. When having 3 LSP clients started (ID 1,
2, 3) and doing `:LspRestart 3`, the 3rd client is stopped and a new
client (with ID 4) is started. The final list of clients is 1, 2, 4.
`ipairs` stops the iteration at 2 and does not iterate over the client
with ID 4. `pairs` handles all clients in the table.

`pairs` is also safe to use because the loop does not rely on the index
of the element.

This was previously fixed in
https://github.com/weilbith/nvim-code-action-menu/pull/5 but looks like
it regressed in
https://github.com/weilbith/nvim-code-action-menu/commit/7361b5637930999d01b93b6ab4c45b784fa0618a#diff-2a5a390abbda6a3c4815db17a50f62a8d78b2567773ed6af509f65365b15745fR107