fix wrong width adaptation of the anchor window
The window stack automatically adapts the width of all windows to the
width of the widest buffer from all windows in the stack. While the
anchor window gets ignored while determining the to use width, it will
also be affected by the width change.
This could be easily fixed by also ignoring the anchor window for the
width change phase.