allow window stack to shrink its width again
The functionality to equalize the width all windows in a stack only
allows to grow the width so far. This was basically the result of the
mistake to use the window width for comparison. Once the stack adapted
its size the first time, it will always just continue to grow by
definition.
The solution to this is to use the buffer width of eachs window in the
stack. This will then also shrink all windows again.