add config for continuous integration pipeline
Having a continuous integration pipeline is very important for most projects. So
it betters gets added earlier than later. For example already now I managed to
have some failing linting. For some reason the git pre-commit hooks has not
worked. There are no checks for the pull requests yet, so this code made it into
the `main` branch.
So far this pipeline only includes a job to lint the code. More checks will get
added in future. Tests are an outstanding topic.