fix broken formatting for Lua module
As described earlier, there was an issue where the formatting was not correct,
the pre-commit hooks was not working and this code made it into the `main`
branch. Now there is this ugly commit without much meaning but a lot of changes.
This sucks when trying to understand code by its git history.