fix diff generation for new files
Some edits returned by the language server create new files. The URI
property points to a path in the filesystem that does not exist yet. As result
the function to read a line from such a file will return `nil`. While this has
been partially handled within the diff generation it wasn't everywhere. The
consequence was an error propagated to the user and a broken menu.