add function to read line from buffer or file
The origin of this issue is that a public function of NeoVim has been removed.
In result the plugin was failing to generate the diff of the code action
changes. To fix this, a custom implementation has been added to read the line of
a file or buffer. The former origin function was taken as inspiration and got
optimized for the specific use-case here.