update CI base image
This was triggered by the issue that my local ArchLinux was already on more
recent version of the `stylua` package then in the currently used CI image
(based on ArchLinux). The versions difference was causing different results
between two pre-commit runs locally and in the CI pipeline.
I must admit that this really not optimal and the version should be rather fixed
in the pre-commit configuration and also installed by it. Unfortunately that
does not work. Or rather doesn't exist yet. I should add this myself at some point.