Use current visual selection whwen asking for range code actions
vim.lsp.util.make_given_range_params uses the **last** visual selection
when asking LSP servers for range actions. This means the actions
returned are outdated, because they are for the previous visual
selection, not the current one.

A workaround to get the right actions is to select a region, <Esc>, and
`gv` to re-select the previous region. Then, the current region will be
the same as the last selection, and the correct range code actions will
be returned.

This commit refactors the logic to use the current visual selection when
asking for range code actions. The workaround is no longer needed.